fixed objects added data info to person object
