+++ renwahsDungeon/src/main/java/cullity/renwahsdungeon/MainApp.java
@@ -1,5 +1,6 @@
 package cullity.renwahsdungeon;
 
+import java.util.ArrayList;
 import javafx.application.Application;
 import static javafx.application.Application.launch;
 import javafx.fxml.FXMLLoader;
@@ -7,15 +8,15 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-
 public class MainApp extends Application {
+
+    public int coins;
+    public ArrayList<Item> inv = new ArrayList();
 
     @Override
     public void start(Stage stage) throws Exception {
         Parent root = FXMLLoader.load(getClass().getResource("/fxml/Scene.fxml"));
-        
         Scene scene = new Scene(root);
-        
         stage.setTitle("Renwah's Dungeon");
         stage.setScene(scene);
         stage.show();
